Created firestoreUtils file for save/delete funcitons, reduced repetition
