+++ src/Map/BuildingMarker.tsx
@@ -3,10 +3,10 @@ import { useCallback } from 'react';
 import { InfoWindow, Marker } from '@react-google-maps/api';
 import IBuilding from '../interfaces/IBuilding';
 import { Button } from 'react-bootstrap';
-import firebase from "../db/firebase";
 import { useAuth } from "../contexts/AuthContext";
 import { useState, useContext } from "react";
 import { ModalContext, ModalState } from "../contexts/ModalContext";
+import { saveBuilding, deleteBuilding } from "../utils/firestoreUtils";
 
 interface IBuildingMarkerProps {
   building: IBuilding,
@@ -35,77 +35,18 @@ export function BuildingMarker(props: IBuildingMarkerProps) {
   const {
     buildingID,
     buildingName,
-    phone,
-    phone2,
-    residentialTargetedArea,
-    totalRestrictedUnits,
-    sedu,
-    studioUnits,
-    oneBedroomUnits, 
-    twoBedroomUnits,
-    threePlusBedroomUnits,
-    urlForBuilding,
-    streetNum,
-    street, 
-    city,
-    state, 
-    zip,
-    lat, 
-    lng,
   } = building;
 
   const [isSaved, setIsSaved] = useState(false);
 
   function toggleSave() {
-    (wasOriginallySaved || isSaved) ? deleteBuilding() : saveBuilding();
-  }
-
-  function saveBuilding() {
-    setIsSaved(true);
-    firebase.firestore().collection("users").doc(currentUser.uid).collection("savedHomes").doc(buildingID)
-    .set({
-      "buildingID": buildingID,
-      "buildingName": buildingName,
-      "phone": phone,
-      "phone2": phone2,
-      "residentialTargetedArea": residentialTargetedArea,
-      "totalRestrictedUnits": totalRestrictedUnits,
-      "sedu": sedu,
-      "studioUnits": studioUnits,
-      "oneBedroomUnits": oneBedroomUnits, 
-      "twoBedroomUnits": twoBedroomUnits,
-      "threePlusBedroomUnits": threePlusBedroomUnits,
-      "urlForBuilding": urlForBuilding,
-      "streetNum": streetNum,
-      "street": street, 
-      "city": city,
-      "state": state, 
-      "zip": zip,
-      "lat": lat,
-      "lng": lng,
-    })
-    .then(() => {
-      console.log(`${buildingName} saved to user list`);
-    })
-    .catch((error) => {
-      console.error("Error adding document: ", error);
-    });
-  };
-
-  function deleteBuilding() {
-    setIsSaved(false);
-    const savedHomesQuery = firebase.firestore().collection("users").doc(currentUser.uid).collection("savedHomes").where('buildingID','==', buildingID);
-    savedHomesQuery.get().then(function(querySnapshot) {
-      querySnapshot.forEach(function(doc) {
-        doc.ref.delete()
-        .then(() => {
-          console.log(`${buildingName} deleted from user list`);
-        })
-        .catch((error) => {
-          console.error("Error deleting document: ", error);
-        });
-      });
-    });
+    if (wasOriginallySaved || isSaved) {
+      setIsSaved(false);
+      deleteBuilding(currentUser, buildingID, buildingName);
+    } else {
+      setIsSaved(true);
+      saveBuilding(currentUser, building);
+    }
   };
 
   return (
