+++ src/components/SavedHomesCard.tsx
@@ -4,6 +4,8 @@ import { useAuth } from "../contexts/AuthContext";
 import { Card, Col, Container, Form, Button, ListGroup, ListGroupItem, Row } from 'react-bootstrap';
 import { XLg} from "react-bootstrap-icons"
 import ISavedBuilding from "../interfaces/ISavedBuilding";
+import { deleteBuilding } from "../utils/firestoreUtils";
+
 
 export function SavedHomesCard(props: ISavedBuilding) {
   const { currentUser } = useAuth() as any
@@ -27,21 +29,6 @@ export function SavedHomesCard(props: ISavedBuilding) {
     zip,
     note
   } = props;
-
-  function deleteBuilding(e: any) {
-    const savedHomesQuery = firebase.firestore().collection("users").doc(currentUser.uid).collection("savedHomes").where('buildingID','==', buildingID);
-    savedHomesQuery.get().then(function(querySnapshot) {
-      querySnapshot.forEach(function(doc) {
-        doc.ref.delete()
-        .then(() => {
-          console.log(`${buildingName} deleted from user list`);
-        })
-        .catch((error) => {
-          console.error("Error deleting document: ", error);
-        });
-      });
-    });
-  };
 
   const [noteToAdd, setNoteToAdd] = useState(note)
 
@@ -92,7 +79,10 @@ export function SavedHomesCard(props: ISavedBuilding) {
                 </Card.Title>
               </Col>
               <Col className="p-0 col-1">
-                <XLg onClick={deleteBuilding} className="card-close-icon"/>
+                <XLg
+                  onClick={() => {deleteBuilding(currentUser, buildingID, buildingName)}}
+                  className="card-close-icon"
+                />
               </Col>
             </Row>
             <Row>
